UISlider和UIScrollbar的OnValueChange事件存在预制体里的是float，但是UE5中改为了double，这样加载的预制体的类型就错了，所以在UE5里强制改为double

+++ Source/LGUI/Private/PlayTween/LGUIPlayTween.cpp
@@ -22,7 +22,7 @@ void ULGUIPlayTween::Start()
 			onStart.FireEvent();
 		})
 		->OnUpdate([&](float progress) {
-			onUpdateProgress.FireEvent((double)progress);
+			onUpdateProgress.FireEvent(progress);
 		})
 		->OnCycleComplete([&] {
 			onCycleComplete.FireEvent();
